requirements.txt
asgiref==3.8.1
Django==4.2.11
djangorestframework==3.15.1
sqlparse==0.5.0
typing-extensions==4.11.0
tzdata==2024.1
